refactor(Extensions): Move all navigation extensions to own file
